refactor auth flow to remove otp signin
