+++ client/src/App.tsx
@@ -10,7 +10,6 @@ import { setUser } from './store/slices/userSlice';
 import { setAdminToken } from './store/slices/adminSlice';
 import type { AppDispatch } from './store';
 import Loader from './components/Loader';
-import { ensureInvisibleRecaptcha } from './lib/firebase';
 
 const Landing = lazy(() => import('./pages/Landing/Landing'));
 const Login = lazy(() => import('./pages/auth/Login/Login'));
@@ -18,7 +17,6 @@ const Signup = lazy(() => import('./pages/auth/Signup/Signup'));
 const OTP = lazy(() => import('./pages/auth/OTP/OTP'));
 const ForgotPassword = lazy(() => import('./pages/auth/ForgotPassword/ForgotPassword'));
 const SetNewPassword = lazy(() => import('./pages/auth/SetNewPassword/SetNewPassword'));
-const PhoneAuth = lazy(() => import('./pages/auth/PhoneAuth/PhoneAuth'));
 const Profile = lazy(() => import('./pages/Profile/Profile'));
 const Home = lazy(() => import('./pages/Home/Home'));
 const Shops = lazy(() => import('./pages/Shops/Shops'));
@@ -66,8 +64,6 @@ function App() {
     if (adminToken) {
       dispatch(setAdminToken(adminToken));
     }
-
-    ensureInvisibleRecaptcha('recaptcha-container');
   }, [dispatch]);
 
   return (
@@ -79,7 +75,6 @@ function App() {
         <Route path="/login" element={<Login />} />
         <Route path="/signup" element={<Signup />} />
         <Route path="/otp" element={<OTP />} />
-        <Route path="/phone-auth" element={<PhoneAuth />} />
         <Route path="/forgot-password" element={<ForgotPassword />} />
         <Route path="/set-new-password" element={<SetNewPassword />} />
         <Route path="/admin/login" element={<AdminLogin />} />
