+++ client/src/pages/auth/ForgotPassword/ForgotPassword.tsx
@@ -1,17 +1,32 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { motion } from 'framer-motion';
+import { RecaptchaVerifier, signInWithPhoneNumber } from 'firebase/auth';
 import './ForgotPassword.scss';
 import Loader from '../../../components/Loader';
 import showToast from '../../../components/ui/Toast';
-import { sendOtpToPhone } from '../../../lib/firebase';
 import { mapFirebaseError } from '../../../lib/firebaseErrors';
+import { auth } from '../../../services/firebase.config';
 
 const ForgotPassword = () => {
   const navigate = useNavigate();
   const [phone, setPhone] = useState('');
   const [error, setError] = useState('');
   const [loading, setLoading] = useState(false);
+
+  const onCaptchVerify = () => {
+    if (!(window as any).recaptchaVerifier) {
+      (window as any).recaptchaVerifier = new RecaptchaVerifier(
+        'recaptcha-container',
+        {
+          size: 'invisible',
+          callback: () => {},
+          'expired-callback': () => {},
+        },
+        auth
+      );
+    }
+  };
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -22,7 +37,10 @@ const ForgotPassword = () => {
     const phoneE164 = `+91${phone}`;
     try {
       setLoading(true);
-      await sendOtpToPhone(phoneE164);
+      onCaptchVerify();
+      const appVerifier = (window as any).recaptchaVerifier;
+      const confirmation = await signInWithPhoneNumber(auth, phoneE164, appVerifier);
+      (window as any).confirmationResult = confirmation;
       showToast(`OTP sent to ${phoneE164}`, 'success');
       navigate(`/otp?purpose=reset&phone=${encodeURIComponent(phoneE164)}`);
     } catch (err: any) {
